refactored some common code needed by the server
